+++ FakeServer/Authentication/Basic/BasicAuthentication.cs
@@ -3,7 +3,10 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Text;
@@ -44,6 +47,31 @@ namespace FakeServer.Authentication.Basic
         {
             builder.Services.TryAddEnumerable(ServiceDescriptor.Singleton<IPostConfigureOptions<BasicTokenOptions>, BasicTokenPostConfigureOptions>());
             return builder.AddScheme<BasicTokenOptions, BasicAuthenticationHandler>(authenticationScheme, displayName, configureOptions);
+        }
+        public static OpenApiSecurityScheme GetBasicSecurityDefinition(this SwaggerGenOptions s)
+        {
+            return new OpenApiSecurityScheme
+            {
+                Type = SecuritySchemeType.Http,
+                Scheme = BasicAuthenticationDefaults.AuthenticationScheme,
+                In = ParameterLocation.Header
+            };
+        }
+        public static OpenApiSecurityRequirement GetBasicSecurityRequirement(this SwaggerGenOptions s)
+        {
+            var securityRequirement = new OpenApiSecurityRequirement();
+            var scheme = new OpenApiSecurityScheme
+            {
+                Reference = new OpenApiReference
+                {
+                    Type = ReferenceType.SecurityScheme,
+                    Id = BasicAuthenticationDefaults.AuthenticationScheme
+                }
+            };
+
+            securityRequirement.Add(scheme, new List<string>());
+
+            return securityRequirement;
         }
     }
